requirements.txt
Flask
opencv-python
pillow
onnxruntime
matplotlib
pyyaml
h5py
tqdm
scipy
pandas